feat: refactor & componetization
